Add sidebar file explorer, strip editor to single-file Monaco viewer
- Add file explorer WebView in the left sidebar with git status, tree
  navigation, context menus, inline rename, and new file/folder creation
- Clicking a file in the sidebar opens it as a new editor tab (one file
  per panel)
- Strip editor panel down to single-file Monaco: no built-in explorer,
  no tabs, no folder watching — just readFile/writeFile/save
- Editor auto-opens its file when Monaco signals ready (editorReady msg)
- Explorer uses its own JS bridge (cmuxExplorer) with self-contained
  file system operations separate from the editor's cmuxEditor bridge
