+++ Resources/explorer/explorer.css
@@ -1,0 +1,204 @@
+/* cmux Sidebar Explorer — file tree only */
+@import url('https://cdn.jsdelivr.net/npm/@vscode/codicons@0.0.36/dist/codicon.css');
+
+* { margin: 0; padding: 0; box-sizing: border-box; }
+
+:root {
+    --sidebar-bg: #181818;
+    --sidebar-fg: #cccccc;
+    --sidebar-border: #2b2b2b;
+    --sidebar-header-bg: #181818;
+    --list-hover-bg: #2a2d2e;
+    --list-active-selection-bg: #094771;
+    --list-active-selection-fg: #ffffff;
+    --list-inactive-selection-bg: #37373d;
+    --list-focus-outline: #007fd4;
+    --tree-indent-guide: #585858;
+    --tree-row-height: 22px;
+    --git-modified: #e2c08d;
+    --git-added: #81b88b;
+    --git-deleted: #c74e39;
+    --git-untracked: #73c991;
+    --git-renamed: #73c991;
+    --git-ignored: #8c8c8c;
+    --git-conflict: #e4676b;
+    --focus-border: #0078d4;
+    --scrollbar-thumb: rgba(121,121,121,0.4);
+    --input-bg: #313131;
+    --input-border: #3c3c3c;
+    --input-fg: #cccccc;
+    --context-menu-bg: #1f1f1f;
+    --context-menu-hover: #094771;
+}
+
+html, body {
+    height: 100%;
+    overflow: hidden;
+    background: var(--sidebar-bg);
+    color: var(--sidebar-fg);
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+    font-size: 13px;
+    -webkit-user-select: none;
+    user-select: none;
+}
+
+#app {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    width: 100%;
+}
+
+#sidebar-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 8px 0 12px;
+    height: 35px;
+    font-size: 11px;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+    color: var(--sidebar-fg);
+    flex-shrink: 0;
+}
+
+.header-buttons { display: flex; gap: 0; }
+
+.header-btn {
+    background: none;
+    border: none;
+    color: var(--sidebar-fg);
+    opacity: 0.7;
+    cursor: pointer;
+    font-family: 'codicon';
+    font-size: 16px;
+    width: 22px;
+    height: 22px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 3px;
+    line-height: 1;
+}
+.header-btn:hover { opacity: 1; background: var(--list-hover-bg); }
+
+#file-tree {
+    flex: 1;
+    overflow-y: auto;
+    overflow-x: hidden;
+}
+
+#file-tree::-webkit-scrollbar { width: 10px; background: transparent; }
+#file-tree::-webkit-scrollbar-thumb { background: var(--scrollbar-thumb); border-radius: 0; min-height: 40px; }
+#file-tree::-webkit-scrollbar-thumb:hover { background: rgba(121,121,121,0.7); }
+
+.tree-row {
+    display: flex;
+    align-items: center;
+    height: var(--tree-row-height);
+    line-height: var(--tree-row-height);
+    cursor: pointer;
+    padding-right: 8px;
+    position: relative;
+    white-space: nowrap;
+}
+.tree-row:hover { background: var(--list-hover-bg); }
+.tree-row.selected { background: var(--list-inactive-selection-bg); }
+.tree-row.selected.focused { background: var(--list-active-selection-bg); color: var(--list-active-selection-fg); }
+
+.tree-indent { display: inline-flex; height: 100%; flex-shrink: 0; }
+.indent-guide { display: inline-block; width: 8px; height: 100%; border-right: 1px solid transparent; box-sizing: border-box; }
+.indent-guide.active { border-right-color: var(--tree-indent-guide); }
+
+.tree-twistie {
+    width: 16px;
+    height: var(--tree-row-height);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-shrink: 0;
+    font-family: 'codicon';
+    font-size: 16px;
+    color: var(--sidebar-fg);
+    transition: transform 0.1s ease;
+}
+.tree-twistie.expanded { transform: rotate(90deg); }
+.tree-twistie.hidden { visibility: hidden; }
+
+.tree-icon {
+    width: 16px; height: 16px;
+    margin-right: 6px; flex-shrink: 0;
+    font-family: 'codicon'; font-size: 16px;
+    display: flex; align-items: center; justify-content: center;
+    line-height: 1;
+}
+
+.tree-icon.icon-folder { color: #dcb67a; }
+.tree-icon.icon-folder-open { color: #dcb67a; }
+.tree-icon.icon-file { color: var(--sidebar-fg); opacity: 0.7; }
+.tree-icon.icon-js { color: #e6cd69; }
+.tree-icon.icon-ts { color: #3178c6; }
+.tree-icon.icon-json { color: #e6cd69; }
+.tree-icon.icon-html { color: #e34c26; }
+.tree-icon.icon-css { color: #563d7c; }
+.tree-icon.icon-md { color: #519aba; }
+.tree-icon.icon-py { color: #3572a5; }
+.tree-icon.icon-rs { color: #dea584; }
+.tree-icon.icon-go { color: #00add8; }
+.tree-icon.icon-swift { color: #f05138; }
+.tree-icon.icon-rb { color: #cc342d; }
+.tree-icon.icon-java { color: #b07219; }
+.tree-icon.icon-c { color: #555555; }
+.tree-icon.icon-cpp { color: #f34b7d; }
+.tree-icon.icon-sh { color: #89e051; }
+.tree-icon.icon-yml { color: #cb171e; }
+.tree-icon.icon-toml { color: #9c4221; }
+.tree-icon.icon-lock { color: #8c8c8c; opacity: 0.5; }
+.tree-icon.icon-image { color: #a074c4; }
+.tree-icon.icon-zig { color: #f7a41d; }
+
+.tree-label { flex: 1; overflow: hidden; text-overflow: ellipsis; font-size: 13px; line-height: var(--tree-row-height); }
+.tree-label.git-modified { color: var(--git-modified); }
+.tree-label.git-added { color: var(--git-added); }
+.tree-label.git-deleted { color: var(--git-deleted); text-decoration: line-through; }
+.tree-label.git-untracked { color: var(--git-untracked); }
+.tree-label.git-renamed { color: var(--git-renamed); }
+.tree-label.git-conflict { color: var(--git-conflict); }
+.tree-label.git-ignored { color: var(--git-ignored); opacity: 0.4; }
+
+.tree-badge { margin-left: auto; padding: 0 5px; font-size: 11px; font-weight: 600; border-radius: 3px; flex-shrink: 0; line-height: 16px; min-width: 16px; text-align: center; }
+.tree-badge.badge-M { color: var(--git-modified); }
+.tree-badge.badge-A { color: var(--git-added); }
+.tree-badge.badge-D { color: var(--git-deleted); }
+.tree-badge.badge-U { color: var(--git-untracked); }
+.tree-badge.badge-R { color: var(--git-renamed); }
+.tree-badge.badge-C { color: var(--git-conflict); }
+.tree-badge.badge-I { color: var(--git-ignored); }
+.tree-badge.badge-dot { width: 6px; height: 6px; min-width: 6px; padding: 0; border-radius: 50%; margin-left: 6px; align-self: center; }
+
+.tree-children { display: none; }
+.tree-children.expanded { display: block; }
+
+.context-menu {
+    position: fixed;
+    background: var(--context-menu-bg);
+    border: 1px solid var(--sidebar-border);
+    border-radius: 4px;
+    padding: 4px 0;
+    min-width: 180px;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.5);
+    z-index: 1000;
+    font-size: 13px;
+}
+.context-menu-item { padding: 4px 24px 4px 12px; cursor: pointer; display: flex; align-items: center; gap: 8px; height: 26px; }
+.context-menu-item:hover { background: var(--context-menu-hover); color: #fff; }
+.context-menu-item .shortcut { margin-left: auto; font-size: 11px; opacity: 0.6; }
+.context-menu-separator { height: 1px; background: var(--sidebar-border); margin: 4px 0; }
+
+.tree-inline-input {
+    width: calc(100% - 8px); margin: 0 4px; padding: 0 4px; height: 20px;
+    background: var(--input-bg); border: 1px solid var(--focus-border); border-radius: 2px;
+    color: var(--input-fg); font-size: 13px; font-family: inherit; outline: none; line-height: 20px;
+}
+.tree-input-row { height: var(--tree-row-height); display: flex; align-items: center; padding: 0 !important; }
