+++ Resources/explorer/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>cmux Explorer</title>
+    <link rel="stylesheet" href="explorer.css">
+</head>
+<body>
+    <div id="app">
+        <div id="sidebar-header">
+            <span id="project-name">EXPLORER</span>
+        </div>
+        <div id="file-tree"></div>
+    </div>
+
+    <script src="explorer.js"></script>
+</body>
+</html>
